--- a/benchmarks/RUN.md
+++ b/benchmarks/RUN.md
@@ -7,6 +7,7 @@ go test -benchmem -run=^$ -bench ^BenchmarkShortestPathManyToMany$ github.com/Ld
 git checkout dcb59c8c6cbac82090b9fcab8ec256d678765a74 && \
 go test -benchmem -run=^$ -bench ^BenchmarkShortestPathManyToMany$ github.com/LdDl/ch -v -count=1 > benchmarks/old_mn_ptr.txt && \
 benchcmp benchmarks/old_mn_ptr.txt benchmarks/new_mn_ptr.txt && \
+benchstat benchmarks/old_mn_ptr.txt benchmarks/new_mn_ptr.txt && \
 git checkout redudant_pointers
 ```
 
@@ -17,6 +18,7 @@ go test -benchmem -run=^$ -bench ^BenchmarkShortestPathOneToMany$ github.com/LdD
 git checkout dcb59c8c6cbac82090b9fcab8ec256d678765a74 && \
 go test -benchmem -run=^$ -bench ^BenchmarkShortestPathOneToMany$ github.com/LdDl/ch -v -count=1 > benchmarks/old_1n_ptr.txt && \
 benchcmp benchmarks/old_1n_ptr.txt benchmarks/new_1n_ptr.txt && \
+benchstat benchmarks/old_1n_ptr.txt benchmarks/new_1n_ptr.txt && \
 git checkout redudant_pointers
 ```
 
@@ -27,6 +29,7 @@ go test -benchmem -run=^$ -bench ^BenchmarkShortestPath$ github.com/LdDl/ch -v -
 git checkout dcb59c8c6cbac82090b9fcab8ec256d678765a74 && \
 go test -benchmem -run=^$ -bench ^BenchmarkShortestPath$ github.com/LdDl/ch -v -count=1 > benchmarks/old_11_ptr.txt && \
 benchcmp benchmarks/old_11_ptr.txt benchmarks/new_11_ptr.txt && \
+benchstat benchmarks/old_11_ptr.txt benchmarks/new_11_ptr.txt && \
 git checkout redudant_pointers
 ```
 
@@ -37,6 +40,7 @@ go test -benchmem -run=^$ -bench ^BenchmarkStaticCaseShortestPath$ github.com/Ld
 git checkout dcb59c8c6cbac82090b9fcab8ec256d678765a74 && \
 go test -benchmem -run=^$ -bench ^BenchmarkStaticCaseShortestPath$ github.com/LdDl/ch -v -count=1 > benchmarks/old_11static_ptr.txt && \
 benchcmp benchmarks/old_11static_ptr.txt benchmarks/new_11static_ptr.txt && \
+benchstat benchmarks/old_11static_ptr.txt benchmarks/new_11static_ptr.txt && \
 git checkout redudant_pointers
 ```
 
@@ -47,6 +51,7 @@ go test -benchmem -run=^$ -bench ^BenchmarkPrepareContracts$ github.com/LdDl/ch 
 git checkout dcb59c8c6cbac82090b9fcab8ec256d678765a74 && \
 go test -benchmem -run=^$ -bench ^BenchmarkPrepareContracts$ github.com/LdDl/ch -v -count=1 > benchmarks/old_ch_prepare_ptr.txt && \
 benchcmp benchmarks/old_ch_prepare_ptr.txt benchmarks/new_ch_prepare_ptr.txt && \
+benchstat benchmarks/old_ch_prepare_ptr.txt benchmarks/new_ch_prepare_ptr.txt && \
 git checkout redudant_pointers
 ```