--- a/graph.go
+++ b/graph.go
@@ -17,7 +17,7 @@ type Graph struct {
 	restrictions map[int64]map[int64]int64
 	mapping      map[int64]int64
 
-	Vertices     []*Vertex
+	Vertices     []Vertex
 	edgesNum     int64
 	shortcutsNum int64
 
@@ -29,7 +29,7 @@ type Graph struct {
 func NewGraph() *Graph {
 	return &Graph{
 		mapping:      make(map[int64]int64),
-		Vertices:     make([]*Vertex, 0),
+		Vertices:     make([]Vertex, 0),
 		edgesNum:     0,
 		shortcutsNum: 0,
 		shortcuts:    make(map[int64]map[int64]*ShortcutPath),
@@ -47,7 +47,7 @@ func (graph *Graph) CreateVertex(label int64) error {
 	if graph.frozen {
 		return ErrGraphIsFrozen
 	}
-	v := &Vertex{
+	v := Vertex{
 		Label:        label,
 		delNeighbors: 0,
 		distance:     NewDistance(),
@@ -142,7 +142,7 @@ func (graph *Graph) computeImportance() *importanceHeap {
 	heap.Init(pqImportance)
 	for i := range graph.Vertices {
 		graph.Vertices[i].computeImportance()
-		heap.Push(pqImportance, graph.Vertices[i])
+		heap.Push(pqImportance, &graph.Vertices[i])
 	}
 	graph.Freeze()
 	return pqImportance