--- a/vertex.go
+++ b/vertex.go
@@ -2,9 +2,9 @@ package ch
 
 // Vertex All information about vertex
 type Vertex struct {
-	distance         *Distance
 	inIncidentEdges  []incidentEdge
 	outIncidentEdges []incidentEdge
+	distance         Distance
 
 	vertexNum int64
 	Label     int64
@@ -86,8 +86,8 @@ type Distance struct {
 }
 
 // NewDistance Constructor for Distance
-func NewDistance() *Distance {
-	return &Distance{
+func NewDistance() Distance {
+	return Distance{
 		previousOrderPos: -1,
 		previousSourceID: -1,
 		distance:         Infinity,